[fix](ABC-722): Move back fetch file to primitiveIcon for Salesforce
