+++ src/modules/base/primitiveIcon/fetch.js
