+++ src/modules/base/primitiveIcon/primitiveIcon.js
@@ -34,7 +34,8 @@ import { LightningElement, api } from 'lwc';
 import { classSet } from 'c/utils';
 import { normalizeString as normalize } from 'c/utilsPrivate';
 import standardTemplate from './primitiveIcon.html';
-import { getName, isValidName, polyfill, hasIconLibrary, getIconLibrary, fetchIconLibrary } from 'c/iconUtils';
+import { getName, isValidName, polyfill } from 'c/iconUtils';
+import { fetchIconLibrary, hasIconLibrary, getIconLibrary } from './fetch';
 
 const ICON_SIZES = {valid:['xx-small', 'x-small', 'small', 'medium', 'large'] , default: 'medium'}
 const ICON_VARIANTS = {valid: ['bare', 'error', 'inverse', 'warning', 'success'], default: ''}
